chore: seperate command bot and avoid global state
