+++ docs/sdk/teamsfx.commandbot._constructor_.md
@@ -1,0 +1,24 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [CommandBot](./teamsfx.commandbot.md) &gt; [(constructor)](./teamsfx.commandbot._constructor_.md)
+
+## CommandBot.(constructor)
+
+> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
+> 
+
+Creates a new instance of the `CommandBot`<!-- -->.
+
+<b>Signature:</b>
+
+```typescript
+constructor(adapter: BotFrameworkAdapter, commands?: TeamsFxBotCommandHandler[]);
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  adapter | BotFrameworkAdapter | The bound <code>BotFrameworkAdapter</code>. |
+|  commands | [TeamsFxBotCommandHandler](./teamsfx.teamsfxbotcommandhandler.md)<!-- -->\[\] | The commands to registered with the command bot. Each command should inplement the interface [TeamsFxBotCommandHandler](./teamsfx.teamsfxbotcommandhandler.md) so that it can be correctly handled by this command bot. |
+
