+++ docs/sdk/teamsfx.commandbot.adapter.md
@@ -1,0 +1,14 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [CommandBot](./teamsfx.commandbot.md) &gt; [adapter](./teamsfx.commandbot.adapter.md)
+
+## CommandBot.adapter property
+
+> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
+> 
+
+<b>Signature:</b>
+
+```typescript
+readonly adapter: BotFrameworkAdapter;
+```
