+++ docs/sdk/teamsfx.commandbot.md
@@ -1,0 +1,52 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [CommandBot](./teamsfx.commandbot.md)
+
+## CommandBot class
+
+> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
+> 
+
+A command bot for receiving commands and sending responses in Teams.
+
+<b>Signature:</b>
+
+```typescript
+export declare class CommandBot 
+```
+
+## Remarks
+
+Ensure each command should ONLY be registered with the command once, otherwise it'll cause unexpected behavior if you regiter the same command more than once.
+
+## Example
+
+You can register your commands through the constructor of the [CommandBot](./teamsfx.commandbot.md)<!-- -->, or use the `registerCommand` and `registerCommands` API to add commands after creating the `CommandBot` instance.
+
+```typescript
+// register through constructor
+const commandBot = new CommandBot(adapter, [ new HelloWorldCommandHandler() ]);
+
+// register through `register*` API
+commandBot.registerCommand(new HelpCommandHandler());
+```
+
+## Constructors
+
+|  Constructor | Modifiers | Description |
+|  --- | --- | --- |
+|  [(constructor)(adapter, commands)](./teamsfx.commandbot._constructor_.md) |  | <b><i>(BETA)</i></b> Creates a new instance of the <code>CommandBot</code>. |
+
+## Properties
+
+|  Property | Modifiers | Type | Description |
+|  --- | --- | --- | --- |
+|  [adapter](./teamsfx.commandbot.adapter.md) |  | BotFrameworkAdapter | <b><i>(BETA)</i></b> |
+
+## Methods
+
+|  Method | Modifiers | Description |
+|  --- | --- | --- |
+|  [registerCommand(command)](./teamsfx.commandbot.registercommand.md) |  | <b><i>(BETA)</i></b> Registers a command into the command bot. |
+|  [registerCommands(commands)](./teamsfx.commandbot.registercommands.md) |  | <b><i>(BETA)</i></b> Registers a command into the command bot. |
+
