+++ docs/sdk/teamsfx.commandbot.registercommand.md
@@ -1,0 +1,27 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [CommandBot](./teamsfx.commandbot.md) &gt; [registerCommand](./teamsfx.commandbot.registercommand.md)
+
+## CommandBot.registerCommand() method
+
+> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
+> 
+
+Registers a command into the command bot.
+
+<b>Signature:</b>
+
+```typescript
+registerCommand(command: TeamsFxBotCommandHandler): void;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  command | [TeamsFxBotCommandHandler](./teamsfx.teamsfxbotcommandhandler.md) | The command to registered. |
+
+<b>Returns:</b>
+
+void
+
