--- docs/sdk/teamsfx.conversationoptions.commandhandlers.md
@@ -1,21 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationOptions](./teamsfx.conversationoptions.md) &gt; [commandHandlers](./teamsfx.conversationoptions.commandhandlers.md)
-
-## ConversationOptions.commandHandlers property
-
-> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
-> 
-
-The command handlers to register with the underlying conversation bot that can process a command and return a response.
-
-<b>Signature:</b>
-
-```typescript
-commandHandlers?: TeamsFxBotCommandHandler[];
-```
-
-## Remarks
-
-If provided, the corresponding handler will be invoked if the bot received a message that matches the command pattern (`string` or `RegExp`<!-- -->) defined in the handler.
-
