--- docs/sdk/teamsfx.conversationoptions.enablenotification.md
@@ -1,16 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationOptions](./teamsfx.conversationoptions.md) &gt; [enableNotification](./teamsfx.conversationoptions.enablenotification.md)
-
-## ConversationOptions.enableNotification property
-
-> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
-> 
-
-A boolean, controlling whether to include the notification feature.
-
-<b>Signature:</b>
-
-```typescript
-enableNotification?: boolean;
-```
