--- docs/sdk/teamsfx.conversationoptions.md
@@ -1,25 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationOptions](./teamsfx.conversationoptions.md)
-
-## ConversationOptions interface
-
-> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
-> 
-
-Options to initialize [ConversationBot](./teamsfx.conversationbot.md)<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-export interface ConversationOptions 
-```
-
-## Properties
-
-|  Property | Type | Description |
-|  --- | --- | --- |
-|  [commandHandlers?](./teamsfx.conversationoptions.commandhandlers.md) | [TeamsFxBotCommandHandler](./teamsfx.teamsfxbotcommandhandler.md)<!-- -->\[\] | <b><i>(BETA)</i></b> <i>(Optional)</i> The command handlers to register with the underlying conversation bot that can process a command and return a response. |
-|  [enableNotification?](./teamsfx.conversationoptions.enablenotification.md) | boolean | <b><i>(BETA)</i></b> <i>(Optional)</i> A boolean, controlling whether to include the notification feature. |
-|  [storage?](./teamsfx.conversationoptions.storage.md) | NotificationTargetStorage | <b><i>(BETA)</i></b> <i>(Optional)</i> An optional storage to persist bot notification connections. |
-
