+++ docs/sdk/teamsfx.notificationbot._constructor_.md
@@ -1,18 +1,18 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationBot](./teamsfx.conversationbot.md) &gt; [initialize](./teamsfx.conversationbot.initialize.md)
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [NotificationBot](./teamsfx.notificationbot.md) &gt; [(constructor)](./teamsfx.notificationbot._constructor_.md)
 
-## ConversationBot.initialize() method
+## NotificationBot.(constructor)
 
 > This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
 > 
 
-Initialize bot notification.
+constructor of the notification bot.
 
 <b>Signature:</b>
 
 ```typescript
-static initialize(adapter: BotFrameworkAdapter, options?: ConversationOptions): void;
+constructor(adapter: BotFrameworkAdapter, options?: NotificationOptions);
 ```
 
 ## Parameters
@@ -20,11 +20,7 @@ static initialize(adapter: BotFrameworkAdapter, options?: ConversationOptions): 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
 |  adapter | BotFrameworkAdapter | the bound <code>BotFrameworkAdapter</code> |
-|  options | [ConversationOptions](./teamsfx.conversationoptions.md) | initialize options |
-
-<b>Returns:</b>
-
-void
+|  options | NotificationOptions | initialize options |
 
 ## Remarks
