+++ docs/sdk/teamsfx.notificationbot.installations.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationBot](./teamsfx.conversationbot.md) &gt; [installations](./teamsfx.conversationbot.installations.md)
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [NotificationBot](./teamsfx.notificationbot.md) &gt; [installations](./teamsfx.notificationbot.installations.md)
 
-## ConversationBot.installations() method
+## NotificationBot.installations() method
 
 > This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
 > 
@@ -12,7 +12,7 @@ Get all targets where the bot is installed.
 <b>Signature:</b>
 
 ```typescript
-static installations(): Promise<TeamsBotInstallation[]>;
+installations(): Promise<TeamsBotInstallation[]>;
 ```
 <b>Returns:</b>
