+++ docs/sdk/teamsfx.notificationbot.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationBot](./teamsfx.conversationbot.md)
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [NotificationBot](./teamsfx.notificationbot.md)
 
-## ConversationBot class
+## NotificationBot class
 
 > This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
 > 
@@ -12,7 +12,7 @@ Provide static utilities for bot conversation, including - send notification to 
 <b>Signature:</b>
 
 ```typescript
-export declare class ConversationBot 
+export declare class NotificationBot 
 ```
 
 ## Example
@@ -21,27 +21,30 @@ Here's an example on how to send notification via Teams Bot.
 
 ```typescript
 // initialize (it's recommended to be called before handling any bot message)
-ConversationBot.initialize(adapter, {
-   enableNotification: true
-});
+const notificationBot = new NotificationBot(adapter);
 
 // get all bot installations and send message
-for (const target of await ConversationBot.installations()) {
+for (const target of await notificationBot.installations()) {
   await target.sendMessage("Hello Notification");
 }
 
 // alternative - send message to all members
-for (const target of await ConversationBot.installations()) {
+for (const target of await notificationBot.installations()) {
   for (const member of await target.members()) {
     await member.sendMessage("Hello Notification");
   }
 }
 ```
 
+## Constructors
+
+|  Constructor | Modifiers | Description |
+|  --- | --- | --- |
+|  [(constructor)(adapter, options)](./teamsfx.notificationbot._constructor_.md) |  | <b><i>(BETA)</i></b> constructor of the notification bot. |
+
 ## Methods
 
 |  Method | Modifiers | Description |
 |  --- | --- | --- |
-|  [initialize(adapter, options)](./teamsfx.conversationbot.initialize.md) | <code>static</code> | <b><i>(BETA)</i></b> Initialize bot notification. |
-|  [installations()](./teamsfx.conversationbot.installations.md) | <code>static</code> | <b><i>(BETA)</i></b> Get all targets where the bot is installed. |
+|  [installations()](./teamsfx.notificationbot.installations.md) |  | <b><i>(BETA)</i></b> Get all targets where the bot is installed. |
