+++ docs/sdk/teamsfx.notificationoptions_2.md
@@ -1,0 +1,23 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [NotificationOptions\_2](./teamsfx.notificationoptions_2.md)
+
+## NotificationOptions\_2 interface
+
+> This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
+> 
+
+Options to initialize [NotificationBot](./teamsfx.notificationbot.md)<!-- -->.
+
+<b>Signature:</b>
+
+```typescript
+export interface NotificationOptions 
+```
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [storage?](./teamsfx.notificationoptions_2.storage.md) | NotificationTargetStorage | <b><i>(BETA)</i></b> <i>(Optional)</i> An optional storage to persist bot notification connections. |
+
