+++ docs/sdk/teamsfx.notificationoptions_2.storage.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [ConversationOptions](./teamsfx.conversationoptions.md) &gt; [storage](./teamsfx.conversationoptions.storage.md)
+[Home](./index.md) &gt; [@microsoft/teamsfx](./teamsfx.md) &gt; [NotificationOptions\_2](./teamsfx.notificationoptions_2.md) &gt; [storage](./teamsfx.notificationoptions_2.storage.md)
 
-## ConversationOptions.storage property
+## NotificationOptions\_2.storage property
 
 > This API is provided as a preview for developers and may change based on feedback that we receive. Do not use this API in a production environment.
 >
